--- src/main/kotlin/com/teamsparta/jangtrello/domain/card/dto/DeleteCardRequest.kt
@@ -1,5 +1,0 @@
-package com.teamsparta.jangtrello.domain.card.dto
-
-data class DeleteCardRequest (
-    var password : String
-)
