+++ src/main/kotlin/com/teamsparta/jangtrello/domain/card/repository/CardRepository.kt
@@ -5,7 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository
 
 interface CardRepository : JpaRepository<Card, Long>, CustomCardRepository {
     //fun findByCardListIdAndId(cardListId: Long, cardId: Long): Card?
+
+    // 유저 id 와 일치하는 모든 카드 가져오기
     fun findAllByUserId(id : Long): List<Card>
-    fun findByUserIdAndId(userId : Long, id : Long) : Card?
+
+    // userId 와 cardID가 일치하는 항목 찾기
+    fun findByUserIdAndId(userId : Long, cardId : Long) : Card?
 
 }
