+++ src/main/kotlin/com/teamsparta/jangtrello/domain/user/dto/UserResponse.kt
@@ -1,7 +1,11 @@
 package com.teamsparta.jangtrello.domain.user.dto
 
+import java.time.LocalDateTime
+
 data class UserResponse(
     val id:Long,
+    val createdAt:LocalDateTime?,
+    val modifiedAt: LocalDateTime?,
     var email:String,
     val nickName:String,
     val role : String,
